App refactor: Introduce CursorContextProvider
Improve App module's read- and compose-ability by moving CursorContext
concerns to a separate provider module.
