+++ src/components/CursorContext.ts
@@ -1,7 +1,7 @@
 import {createContext} from 'react';
 import {type CursorPosition} from '../log-update.js';
 
-export type Props = {
+export type CursorContextValue = {
 	/**
 	Set the cursor position relative to the Ink output.
 
@@ -11,7 +11,7 @@ export type Props = {
 };
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-const CursorContext = createContext<Props>({
+const CursorContext = createContext<CursorContextValue>({
 	setCursorPosition() {},
 });
